better proxy errors

+++ Jetty/src/org/mortbay/jetty/servlet/Default.java
@@ -213,11 +213,11 @@ public class Default extends HttpServlet
                 method.equals(HttpRequest.__POST) ||
                 method.equals(HttpRequest.__HEAD))
                 handleGet(request, response,pathInContext,resource,endsWithSlash);   
-            else if (method.equals(HttpRequest.__PUT))
+            else if (_putAllowed && method.equals(HttpRequest.__PUT))
                 handlePut(request, response, pathInContext, resource); 
-            else if (method.equals(HttpRequest.__DELETE))
+            else if (_delAllowed && method.equals(HttpRequest.__DELETE))
                 handleDelete(request, response, pathInContext, resource); 
-            else if (method.equals(HttpRequest.__MOVE))
+            else if (_putAllowed && _delAllowed && method.equals(HttpRequest.__MOVE))
                 handleMove(request, response, pathInContext, resource);
             else if (method.equals(HttpRequest.__OPTIONS))
                 handleOptions(request, response);
